requirements.txt
absl-py==1.4.0
alembic==1.16.5
astunparse==1.6.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
dm-tree==0.1.9
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.1
h5py==3.14.0
idna==3.10
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
six==1.17.0
SQLAlchemy==2.0.43
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-model-optimization==0.8.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
